requirements.txt
altgraph==0.17.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
oauth2client==4.1.3
oauthlib==3.2.0
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pyparsing==3.0.9
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.11
